cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tes-cli"
version = "1.4.5"
description = "TES CLI - Tools"
readme = "README.md"
requires-python = ">=3.14"
license = {text = "MIT"}
authors = [
    {name = "tes4all"}
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.28.0",
    "boto3>=1.42.0",
]

[project.scripts]
tes = "cli.cli:main"

[project.urls]
Homepage = "https://github.com/tes4all/tes"
Repository = "https://github.com/tes4all/tes"

[tool.hatch.build.targets.wheel]
packages = ["src/cli"]
